Handle a redundant cast

+++ mandelbrot/mandelbrot_color_functions.h
@@ -70,15 +70,7 @@
 
       using std::sin;
 
-      const float
-        my_color
-        {
-          (std::max)
-          (
-            static_cast<float>((sin(color_phase) / 2.0F) + 0.5F),
-            0.0F
-          )
-        };
+      const float my_color { (std::max)((sin(color_phase) / 2.0F) + 0.5F, 0.0F) };
 
       return static_cast<std::uint_fast32_t>(float { my_color * 255.0F }); // NOLINT(cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers)
     }
